+++ src/main/java/com/intelliFitHub/fitness/common/user/service/CalculationService.java
@@ -3,5 +3,5 @@ package com.intelliFitHub.fitness.common.user.service;
 import com.intelliFitHub.fitness.common.dao.domain.TdeeCalculator;
 
 public interface CalculationService {
-    double CalculateTDEE(TdeeCalculator tdeeCalculator) throws Exception;
+    int CalculateTDEE(TdeeCalculator tdeeCalculator) throws Exception;
 }
